ExtremeWorldBackend/src/test/java/extremeworld/controller/UserControllerTest.java
package extremeworld.controller;

import Util.UserDtoUtil;
import com.fasterxml.jackson.databind.ObjectMapper;
import extremeworld.dto.UserDTO;
import extremeworld.service.UserService;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.MvcResult;

import static org.assertj.core.api.Assertions.assertThat;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

import static org.junit.jupiter.api.Assertions.*;

@WebMvcTest(controllers = UserController.class)
class UserControllerTest {

    @MockBean
    private UserService userService;

    @Autowired
    private MockMvc mockMvc;

    @Autowired
    private ObjectMapper objectMapper;

    @Test
    @DisplayName("CreateUser_Test")
    void createUser() throws Exception {
        UserDTO userDTO = UserDtoUtil.generateUserDTO("Test", "User");
        when(userService.createUser(any(), any())).thenReturn(userDTO);

        MvcResult result = mockMvc.perform(post("/users/client")
                .content(objectMapper.writeValueAsString(userDTO))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andReturn();

        assertThat(result.getResponse().getContentAsString()).isEqualTo(objectMapper.writeValueAsString(userDTO));
    }

    @Test
    @DisplayName("GetUserById_Test")
    void getUserById() throws Exception{


    }

    @Test
    void deleteOne() {
    }

    @Test
    void updateStudent() {
    }
}